Refactore useIsMounted to separate hook file
